Relocate exceptions
