+++ src/core/Account/domain/exceptions/AccountNotFoundError.ts
@@ -1,4 +1,4 @@
-import { AccountId } from '../domain/value-object/AccountId';
+import { AccountId } from '../value-object/AccountId';
 
 export class AccountNotFoundError extends Error {
   constructor(accountId: AccountId) {
